+++ app/src/main/java/co/kr/woowahan_repo/di/ServiceLocator.kt
@@ -2,7 +2,7 @@ package co.kr.woowahan_repo.di
 
 import co.kr.woowahan_repo.BuildConfig
 import co.kr.woowahan_repo.data.api.interceptor.AuthInterceptor
-import co.kr.woowahan_repo.data.model.repositoryimpl.GithubRepositorySearchRepositoryImpl
+import co.kr.woowahan_repo.data.repository.GithubRepositorySearchRepositoryImpl
 import co.kr.woowahan_repo.data.repository.GithubIssuesRepositoryImpl
 import co.kr.woowahan_repo.data.service.GithubOAuthAccessTokenService
 import co.kr.woowahan_repo.data.service.NotificationsService
